move agEUR redemption to tip_118 proposal
