+++ proposals/dao/deprecate_incentives.ts
@@ -5,11 +5,10 @@ import {
   NamedAddresses,
   SetupUpgradeFunc,
   TeardownUpgradeFunc,
-  PcvStats,
   ValidateUpgradeFunc
 } from '@custom-types/types';
 import { BigNumber } from 'ethers';
-import { getImpersonatedSigner, overwriteChainlinkAggregator } from '@test/helpers';
+import { getImpersonatedSigner } from '@test/helpers';
 import { forceEth } from '@test/integration/setup/utils';
 
 /*
@@ -22,20 +21,13 @@ const fipNumber = 'TIP-114: Deprecate TRIBE Incentives system';
 
 let initialCoreTribeBalance: BigNumber;
 let initialRariDelegatorBalance: BigNumber;
-let pcvStatsBefore: PcvStats;
-let daiBalanceBefore: BigNumber;
 
 // Do any deployments
 // This should exclusively include new contract deployments
 const deploy: DeployUpgradeFunc = async (deployAddress: string, addresses: NamedAddresses, logging: boolean) => {
-  // Deploy agEUR Redeemer contract
-  const angleEuroRedeemerFactory = await ethers.getContractFactory('AngleEuroRedeemer');
-  const angleEuroRedeemer = await angleEuroRedeemerFactory.deploy();
-  await angleEuroRedeemer.deployed();
-  logging && console.log(`angleEuroRedeemer: ${angleEuroRedeemer.address}`);
-
+  console.log(`No deploy actions for fip${fipNumber}`);
   return {
-    angleEuroRedeemer
+    // put returned contract objects here
   };
 };
 
@@ -43,30 +35,10 @@ const deploy: DeployUpgradeFunc = async (deployAddress: string, addresses: Named
 // This could include setting up Hardhat to impersonate accounts,
 // ensuring contracts have a specific state, etc.
 const setup: SetupUpgradeFunc = async (addresses, oldContracts, contracts, logging) => {
-  // deprecate incentives setup
   const tribe = contracts.tribe;
+
   initialCoreTribeBalance = await tribe.balanceOf(addresses.core);
   initialRariDelegatorBalance = await tribe.balanceOf(addresses.rariRewardsDistributorDelegator);
-
-  // make sure ETH oracle is fresh (for B.AMM not to revert, etc)
-  // Read Chainlink ETHUSD price & override chainlink storage to make it a fresh value
-  const ethPrice = (await contracts.chainlinkEthUsdOracleWrapper.read())[0].toString() / 1e10;
-  await overwriteChainlinkAggregator(addresses.chainlinkEthUsdOracle, Math.round(ethPrice).toString(), '8');
-
-  // read pcvStats before proposal execution
-  pcvStatsBefore = await contracts.collateralizationOracle.pcvStats();
-
-  // angle multisig action : make enough USDC collateral available for redemptions
-  const angleMultisigSigner = await getImpersonatedSigner('0x0C2553e4B9dFA9f83b1A6D3EAB96c4bAaB42d430');
-  await forceEth(angleMultisigSigner.address);
-  await contracts.anglePoolManagerUsdc.connect(angleMultisigSigner).updateStrategyDebtRatio(
-    addresses.angleStrategyUsdc1, // USDC strategy has 57M deployed
-    '0'
-  );
-  await contracts.angleStrategyUsdc1.harvest();
-
-  // read DAI PSM balance before proposal execution
-  daiBalanceBefore = await contracts.dai.balanceOf(addresses.daiFixedPricePSM);
 };
 
 // Tears down any changes made in setup() that need to be
@@ -78,9 +50,6 @@ const teardown: TeardownUpgradeFunc = async (addresses, oldContracts, contracts,
 // Run any validations required on the fip using mocha or console logging
 // IE check balances, check state of contracts, etc.
 const validate: ValidateUpgradeFunc = async (addresses, oldContracts, contracts, logging) => {
-  // ------------------------------------------------------------
-  // Deprecate incentives validation
-  // ------------------------------------------------------------
   const tribe = contracts.tribe;
   const core = contracts.core;
   const tribalChief = contracts.tribalChief;
@@ -145,53 +114,6 @@ const validate: ValidateUpgradeFunc = async (addresses, oldContracts, contracts,
   // 6. Validate TRIBAL_CHIEF_ADMIN_ROLE is revoked
   expect(await core.hasRole(ethers.utils.id('TRIBAL_CHIEF_ADMIN_ROLE'), addresses.tribalCouncilTimelock)).to.be.false;
   expect(await core.hasRole(ethers.utils.id('TRIBAL_CHIEF_ADMIN_ROLE'), addresses.optimisticTimelock)).to.be.false;
-
-  // ------------------------------------------------------------
-  // ANGLE and agEUR Deprecation
-  // ------------------------------------------------------------
-  // check ANGLE token movement
-  expect(await contracts.angle.balanceOf(addresses.angleDelegatorPCVDeposit)).to.be.equal('0');
-  expect(await contracts.angle.balanceOf(addresses.tribalCouncilSafe)).to.be.at.least(
-    ethers.utils.parseEther('200000')
-  );
-  // check deposit is empty
-  expect(await contracts.agEurUniswapPCVDeposit.balance()).to.be.equal('0');
-  expect(await contracts.fei.balanceOf(addresses.agEurUniswapPCVDeposit)).to.be.equal('0');
-  expect(await contracts.agEUR.balanceOf(addresses.agEurUniswapPCVDeposit)).to.be.equal('0');
-
-  // check redemptions of agEUR > DAI
-  const daiRedeemed = (await contracts.dai.balanceOf(addresses.daiFixedPricePSM)).sub(daiBalanceBefore);
-  console.log('daiRedeemed', daiRedeemed.toString() / 1e18);
-  expect(daiRedeemed).to.be.at.least(ethers.utils.parseEther('9500000')); // >9.5M DAI
-
-  // check redeemer is empty
-  expect(await contracts.agEUR.balanceOf(addresses.angleEuroRedeemer)).to.be.equal('0');
-  expect(await contracts.fei.balanceOf(addresses.angleEuroRedeemer)).to.be.equal('0');
-  expect(await contracts.dai.balanceOf(addresses.angleEuroRedeemer)).to.be.equal('0');
-  expect(await contracts.usdc.balanceOf(addresses.angleEuroRedeemer)).to.be.equal('0');
-
-  // display pcvStats
-  console.log('----------------------------------------------------');
-  console.log(' pcvStatsBefore.protocolControlledValue [M]e18 ', Number(pcvStatsBefore.protocolControlledValue) / 1e24);
-  console.log(' pcvStatsBefore.userCirculatingFei      [M]e18 ', Number(pcvStatsBefore.userCirculatingFei) / 1e24);
-  console.log(' pcvStatsBefore.protocolEquity          [M]e18 ', Number(pcvStatsBefore.protocolEquity) / 1e24);
-  const pcvStatsAfter: PcvStats = await contracts.collateralizationOracle.pcvStats();
-  console.log('----------------------------------------------------');
-  console.log(' pcvStatsAfter.protocolControlledValue  [M]e18 ', Number(pcvStatsAfter.protocolControlledValue) / 1e24);
-  console.log(' pcvStatsAfter.userCirculatingFei       [M]e18 ', Number(pcvStatsAfter.userCirculatingFei) / 1e24);
-  console.log(' pcvStatsAfter.protocolEquity           [M]e18 ', Number(pcvStatsAfter.protocolEquity) / 1e24);
-  console.log('----------------------------------------------------');
-  const pcvDiff = pcvStatsAfter.protocolControlledValue.sub(pcvStatsBefore.protocolControlledValue);
-  const cFeiDiff = pcvStatsAfter.userCirculatingFei.sub(pcvStatsBefore.userCirculatingFei);
-  const eqDiff = pcvStatsAfter.protocolEquity.sub(pcvStatsBefore.protocolEquity);
-  console.log(' PCV diff                               [M]e18 ', Number(pcvDiff) / 1e24);
-  console.log(' Circ FEI diff                          [M]e18 ', Number(cFeiDiff) / 1e24);
-  console.log(' Equity diff                            [M]e18 ', Number(eqDiff) / 1e24);
-  console.log('----------------------------------------------------');
-
-  // PCV Equity change should be neutral for this proposal
-  expect(Number(eqDiff) / 1e18).to.be.at.least(-10000);
-  expect(Number(eqDiff) / 1e18).to.be.at.most(+10000);
 };
 
 export { deploy, setup, teardown, validate };
