+++ proposals/dao/tip_118.ts
@@ -6,9 +6,10 @@ import {
   NamedContracts,
   SetupUpgradeFunc,
   TeardownUpgradeFunc,
+  PcvStats,
   ValidateUpgradeFunc
 } from '@custom-types/types';
-import { getImpersonatedSigner } from '@test/helpers';
+import { getImpersonatedSigner, overwriteChainlinkAggregator } from '@test/helpers';
 import { forceEth } from '@test/integration/setup/utils';
 import { BigNumber } from 'ethers';
 import { ERC20HoldingPCVDeposit } from '@custom-types/contracts';
@@ -28,6 +29,8 @@ const toBN = BigNumber.from;
 const fipNumber = 'tip_118';
 
 let initialDAIPSMFeiBalance: BigNumber;
+let pcvStatsBefore: PcvStats;
+let daiBalanceBefore: BigNumber;
 
 // Do any deployments
 // This should exclusively include new contract deployments
@@ -54,7 +57,14 @@ const deploy: DeployUpgradeFunc = async (deployAddress: string, addresses: Named
   await gOHMHoldingPCVDeposit.deployTransaction.wait();
   logging && console.log('gOHM holding deposit deployed to: ', gOHMHoldingPCVDeposit.address);
 
+  // Deploy agEUR Redeemer contract
+  const angleEuroRedeemerFactory = await ethers.getContractFactory('AngleEuroRedeemer');
+  const angleEuroRedeemer = await angleEuroRedeemerFactory.deploy(addresses.core);
+  await angleEuroRedeemer.deployed();
+  logging && console.log(`angleEuroRedeemer: ${angleEuroRedeemer.address}`);
+
   return {
+    angleEuroRedeemer,
     wethHoldingDeposit,
     lusdHoldingDeposit,
     voltHoldingDeposit,
@@ -68,6 +78,26 @@ const deploy: DeployUpgradeFunc = async (deployAddress: string, addresses: Named
 // ensuring contracts have a specific state, etc.
 const setup: SetupUpgradeFunc = async (addresses, oldContracts, contracts, logging) => {
   initialDAIPSMFeiBalance = await contracts.fei.balanceOf(addresses.daiFixedPricePSM);
+
+  // make sure ETH oracle is fresh (for B.AMM not to revert, etc)
+  // Read Chainlink ETHUSD price & override chainlink storage to make it a fresh value
+  const ethPrice = (await contracts.chainlinkEthUsdOracleWrapper.read())[0].toString() / 1e10;
+  await overwriteChainlinkAggregator(addresses.chainlinkEthUsdOracle, Math.round(ethPrice).toString(), '8');
+
+  // read pcvStats before proposal execution
+  pcvStatsBefore = await contracts.collateralizationOracle.pcvStats();
+
+  // angle multisig action : make enough USDC collateral available for redemptions
+  const angleMultisigSigner = await getImpersonatedSigner('0x0C2553e4B9dFA9f83b1A6D3EAB96c4bAaB42d430');
+  await forceEth(angleMultisigSigner.address);
+  await contracts.anglePoolManagerUsdc.connect(angleMultisigSigner).updateStrategyDebtRatio(
+    addresses.angleStrategyUsdc1, // USDC strategy has 57M deployed
+    '0'
+  );
+  await contracts.angleStrategyUsdc1.harvest();
+
+  // read DAI PSM balance before proposal execution
+  daiBalanceBefore = await contracts.dai.balanceOf(addresses.daiFixedPricePSM);
 };
 
 // Tears down any changes made in setup() that need to be
@@ -171,6 +201,53 @@ const validate: ValidateUpgradeFunc = async (addresses, oldContracts, contracts,
 
   // 11. gOHM received funds
   expect(await gOHM.balanceOf(addresses.gOHMHoldingPCVDeposit)).to.be.equal('577180000000000000000');
+
+  // ------------------------------------------------------------
+  // ANGLE and agEUR Deprecation
+  // ------------------------------------------------------------
+  // check ANGLE token movement
+  expect(await contracts.angle.balanceOf(addresses.angleDelegatorPCVDeposit)).to.be.equal('0');
+  expect(await contracts.angle.balanceOf(addresses.tribalCouncilSafe)).to.be.at.least(
+    ethers.utils.parseEther('200000')
+  );
+  // check deposit is empty
+  expect(await contracts.agEurUniswapPCVDeposit.balance()).to.be.equal('0');
+  expect(await contracts.fei.balanceOf(addresses.agEurUniswapPCVDeposit)).to.be.equal('0');
+  expect(await contracts.agEUR.balanceOf(addresses.agEurUniswapPCVDeposit)).to.be.equal('0');
+
+  // check redemptions of agEUR > DAI
+  const daiRedeemed = (await contracts.dai.balanceOf(addresses.daiFixedPricePSM)).sub(daiBalanceBefore);
+  console.log('daiRedeemed', daiRedeemed.toString() / 1e18);
+  expect(daiRedeemed).to.be.at.least(ethers.utils.parseEther('9500000')); // >9.5M DAI
+
+  // check redeemer is empty
+  expect(await contracts.agEUR.balanceOf(addresses.angleEuroRedeemer)).to.be.equal('0');
+  expect(await contracts.fei.balanceOf(addresses.angleEuroRedeemer)).to.be.equal('0');
+  expect(await contracts.dai.balanceOf(addresses.angleEuroRedeemer)).to.be.equal('0');
+  expect(await contracts.usdc.balanceOf(addresses.angleEuroRedeemer)).to.be.equal('0');
+
+  // display pcvStats
+  console.log('----------------------------------------------------');
+  console.log(' pcvStatsBefore.protocolControlledValue [M]e18 ', Number(pcvStatsBefore.protocolControlledValue) / 1e24);
+  console.log(' pcvStatsBefore.userCirculatingFei      [M]e18 ', Number(pcvStatsBefore.userCirculatingFei) / 1e24);
+  console.log(' pcvStatsBefore.protocolEquity          [M]e18 ', Number(pcvStatsBefore.protocolEquity) / 1e24);
+  const pcvStatsAfter: PcvStats = await contracts.collateralizationOracle.pcvStats();
+  console.log('----------------------------------------------------');
+  console.log(' pcvStatsAfter.protocolControlledValue  [M]e18 ', Number(pcvStatsAfter.protocolControlledValue) / 1e24);
+  console.log(' pcvStatsAfter.userCirculatingFei       [M]e18 ', Number(pcvStatsAfter.userCirculatingFei) / 1e24);
+  console.log(' pcvStatsAfter.protocolEquity           [M]e18 ', Number(pcvStatsAfter.protocolEquity) / 1e24);
+  console.log('----------------------------------------------------');
+  const pcvDiff = pcvStatsAfter.protocolControlledValue.sub(pcvStatsBefore.protocolControlledValue);
+  const cFeiDiff = pcvStatsAfter.userCirculatingFei.sub(pcvStatsBefore.userCirculatingFei);
+  const eqDiff = pcvStatsAfter.protocolEquity.sub(pcvStatsBefore.protocolEquity);
+  console.log(' PCV diff                               [M]e18 ', Number(pcvDiff) / 1e24);
+  console.log(' Circ FEI diff                          [M]e18 ', Number(cFeiDiff) / 1e24);
+  console.log(' Equity diff                            [M]e18 ', Number(eqDiff) / 1e24);
+  console.log('----------------------------------------------------');
+
+  // PCV Equity change should be neutral for this proposal
+  expect(Number(eqDiff) / 1e18).to.be.at.least(-10000);
+  expect(Number(eqDiff) / 1e18).to.be.at.most(+10000);
 };
 
 const validateHoldingDepositWithdrawal = async (contracts: NamedContracts, addresses: NamedAddresses) => {
