+++ proposals/description/deprecate_incentives.ts
@@ -2,7 +2,7 @@ import { ethers } from 'ethers';
 import { TemplatedProposalDescription } from '@custom-types/types';
 
 const deprecate_incentives: TemplatedProposalDescription = {
-  title: 'TIP-118: Cleanup part 2',
+  title: 'TIP-114: Deprecate TRIBE Incentives system',
   commands: [
     // Withdraw excess TRIBE from reward system
     {
@@ -84,108 +84,9 @@ const deprecate_incentives: TemplatedProposalDescription = {
       Transfer the admin of the Aave Tribe Incentives Controller Proxy to Aave governance, specifically 
       the LendingPoolAddressesProvider.
       `
-    },
-
-    //// agEUR Redemption
-    {
-      target: 'angleDelegatorPCVDeposit',
-      values: '0',
-      method: 'claimGaugeRewards(address)',
-      arguments: (addresses) => [
-        addresses.angleGaugeUniswapV2FeiAgEur // gauge
-      ],
-      description: 'Claim all unclaimed ANGLE rewards'
-    },
-    {
-      target: 'ratioPCVControllerV2',
-      values: '0',
-      method: 'withdrawRatioERC20(address,address,address,uint256)',
-      arguments: (addresses) => [
-        addresses.angleDelegatorPCVDeposit, // pcvDeposit
-        addresses.angle, // token
-        addresses.tribalCouncilSafe, // to
-        '10000' // basisPoints, 100%
-      ],
-      description: 'Migrate all ANGLE to TC Multisig where it can be sold'
-    },
-    {
-      target: 'angleDelegatorPCVDeposit',
-      values: '0',
-      method: 'unstakeFromGauge(address,uint256)',
-      arguments: (addresses) => [
-        addresses.angleAgEurFeiPool, // token
-        '9295203165522303785936702' // all LP tokens
-      ],
-      description: 'Unstake all agEUR/FEI Uniswap tokens from gauge'
-    },
-    {
-      target: 'ratioPCVControllerV2',
-      values: '0',
-      method: 'withdrawRatioERC20(address,address,address,uint256)',
-      arguments: (addresses) => [
-        addresses.angleDelegatorPCVDeposit, // pcvDeposit
-        addresses.angleAgEurFeiPool, // token
-        addresses.agEurUniswapPCVDeposit, // to
-        '10000' // basisPoints, 100%
-      ],
-      description: 'Move all agEUR/FEI Uniswap tokens to PCV deposit'
-    },
-    {
-      target: 'ratioPCVControllerV2',
-      values: '0',
-      method: 'withdrawRatio(address,address,uint256)',
-      arguments: (addresses) => [
-        addresses.agEurUniswapPCVDeposit, // pcvDeposit
-        addresses.agEurUniswapPCVDeposit, // to
-        '10000' // basisPoints, 100%
-      ],
-      description: 'Remove all agEUR/FEI Uniswap liquidity & burn FEI'
-    },
-    {
-      target: 'ratioPCVControllerV2',
-      values: '0',
-      method: 'withdrawRatioERC20(address,address,address,uint256)',
-      arguments: (addresses) => [
-        addresses.agEurUniswapPCVDeposit, // pcvDeposit
-        addresses.agEUR, // token
-        addresses.angleEuroRedeemer, // to
-        '10000' // basisPoints, 100%
-      ],
-      description: 'Move all agEUR to redeemer'
-    },
-    {
-      target: 'angleEuroRedeemer',
-      values: '0',
-      method: 'redeem()',
-      arguments: (addresses) => [],
-      description: 'Redeem all agEUR for FEI and DAI and send to DAI PSM'
-    },
-    {
-      target: 'collateralizationOracle',
-      values: '0',
-      method: 'removeDeposits(address[])',
-      arguments: (addresses) => [[addresses.agEurUniswapPCVDeposit, addresses.uniswapLensAgEurUniswapGauge]],
-      description: 'Remove agEUR addresses from CR oracle'
-    },
-    {
-      target: 'pcvGuardian',
-      values: '0',
-      method: 'unsetSafeAddress(address)',
-      arguments: (addresses) => [addresses.agEurUniswapPCVDeposit],
-      description: 'Remove agEurUniswapPCVDeposit from safe addresses'
     }
   ],
   description: `
-  TIP-118: Cleanup part 2
-  
-  This vote contains multiple cleanup actions bundled in one DAO vote :
-  - Incentives cleanup (reclaim TRIBE tokens that are not owed to any users)
-  - agEUR redeem (redeem agEUR for USDC and convert to DAI using Maker PSM)
-
-  The details are provided below.
-
-  ----------------------------------------------------------
-
   TIP-114: Deprecate TRIBE Incentives system
 
   Deprecates the TRIBE incentives system according to proposal: 
@@ -196,20 +97,6 @@ const deprecate_incentives: TemplatedProposalDescription = {
   - Withdraws remaining TRIBE from ERC20 Dripper
   - Revokes no longer needed TRIBAL_CHIEF_ADMIN_ROLE roles
   - Transfers the admin of the Aave Fei Incentives Controller Proxy to Aave Governance
-
-  ----------------------------------------------------------
-
-  TIP-110: agEUR & Angle redeem
-
-  Convert all agEUR in the PCV to DAI, and deprecate all Angle Protocol-related contracts, except the contract that is holding vote-escrowed ANGLE.
-
-  This proposal will perform the following actions :
-  - Move all ANGLE rewards to TC Multisig where they can be sold
-  - Unstake from Angle Protocol's Gauge all agEUR/FEI Uniswap tokens
-  - Move all agEUR/FEI Uniswap tokens to PCV deposit
-  - Remove all agEUR/FEI Uniswap liquidity & burn FEI
-  - Redeem agEUR for DAI, and send proceeds to DAI PSM
-  - Unset agEUR safe addresses & CR oracle entries
   `
 };
